Dashboard: fix administration and fpr tests
- Avoid database connection on `fpr.forms` module import. The connection
  was created to get the choices of some fields. However those choices
  are overwritten later in the forms constructor.
- Split, rename and move tests modules to be picked by `pytest`.
- Add `__init__` module to test packages to be picked by Django's runner.
- Change pytest ignore rules in tox.ini.
