+++ tox.ini
@@ -24,7 +24,7 @@ commands =
     pip install -q -r src/archivematicaCommon/requirements/test.txt
     pip install -q -r src/dashboard/src/requirements/test.txt
     pip install -q -r src/MCPServer/requirements/test.txt
-    py.test --ignore=src/dashboard/tests --ignore=src/MCPClient/tests --ignore=src/archivematicaCommon/tests {posargs}
+    py.test --ignore=src/dashboard --ignore=src/MCPClient --ignore=src/archivematicaCommon {posargs}
 
 [testenv:linting]
 basepython = python3
